factor out study configuration constants into a separate file
